Implement support for grizzly and quantum with linuxbridge.
This patch adds support for quantum with linuxbridge. New configuration
setting is being introduced in openstackquickstartrc:

floatingnet=172.31.0.0/29

Which is external/floating network for quantum environment.

Veth devices (vefr,vefq) are used to connect integration bridge with rest of the network (for possible multinode deployment).

On openSUSE 12.3 iproute2 package support for netns is broken, fix will be soon present in update repository (in the mean time script adds security/netfilter repo).

Tested on openSUSE 12.2 and openSUSE 12.2.