Add Sahara
Allow to deploy OpenStack Sahara.

Note: Sahara is still not correctly configured and the services are not
started. But it is still useful to simplifiy testing of Sahara because the
DB and the endpoint are already there.

(cherry picked from commit b1b5fcbe66d61e70f35fb06da49a7f32052d3594)