Add Gnocchi
Allow to deploy OpenStack Gnocchi.

Note: Gnocchi is still not correctly configured and the services
are not started. But it is still useful to simplifiy testing of
Gnocchi because the DB and the endpoint are already there.

(cherry picked from commit 6475d97364c95afe88d8bcc3d04fe98990eb23b9)