Update scripts for package changes
The openstack-nova package is changing the port that the nova placement
API runs on as well as defaulting to enabling the apache vhost that
serves[1] it in order to be compliant with the OpenStack install guide
as well as other package maintainers. This patch updates the quickstart
scripts to sync up with the package changes.

[1] https://build.opensuse.org/request/show/485040