Update role assignment list command
Without this patch, the script emits deprecation warnings about using
the openstackclient 'role list' command, which causes noise in the logs.
This change updates the role queries to use the 'role assignment list'
command and also adds the --format option so that we don't have to parse
the prettytable output. Also updates the group role assignment list
queries to consistently use the long-format flag names.