Update keystone config options
Config options named after keystone 'tenants' were removed[1] and replace
with options named after keystone 'projects'. The result of not setting
the 'project' options was that tempest auth requests would be 'unscoped'
and by default not return a catalog[2], which meant that it could not
look up service endpoints and would fail with

  EndpointNotFound: Endpoint not found
  Details: No matching service found in the catalog.

even though the endpoint was in fact in the catalog. This change fixes
the issue by updating the tempest config to use real config options.
Also remove the compute-admin group which was also dropped[3] as well as
other dropped identity group options[4].

[1] https://review.opendev.org/661664
[2] https://docs.openstack.org/api-ref/identity/v3/?expanded=password-authentication-with-unscoped-authorization-detail#password-authentication-with-unscoped-authorization
[3] https://opendev.org/openstack/tempest/commit/daadcdbec6c8fec4a5e42b4924f9b8480ac222e5
[4] https://opendev.org/openstack/tempest/commit/40847ac3089d02e6b80f370d92762d7f4e15f7b8